refactor(cli): move `EvmArgs` to foundry_cli
