+++ crates/cli/src/opts/mod.rs
@@ -1,6 +1,7 @@
 mod build;
 mod chain;
 mod dependency;
+mod evm;
 mod global;
 mod rpc;
 mod transaction;
@@ -8,6 +9,7 @@ mod transaction;
 pub use build::*;
 pub use chain::*;
 pub use dependency::*;
+pub use evm::*;
 pub use global::*;
 pub use rpc::*;
 pub use transaction::*;
