+++ crates/forge/src/cmd/config.rs
@@ -1,8 +1,8 @@
 use super::build::BuildArgs;
 use clap::Parser;
 use eyre::Result;
-use foundry_cli::utils::LoadConfig;
-use foundry_common::{evm::EvmArgs, shell};
+use foundry_cli::{opts::EvmArgs, utils::LoadConfig};
+use foundry_common::shell;
 use foundry_config::fix::fix_tomls;
 
 foundry_config::impl_figment_convert!(ConfigArgs, build, evm);
