+++ crates/forge/src/cmd/test/mod.rs
@@ -17,12 +17,10 @@ use chrono::Utc;
 use clap::{Parser, ValueHint};
 use eyre::{Context, OptionExt, Result, bail};
 use foundry_cli::{
-    opts::{BuildOpts, GlobalArgs},
+    opts::{BuildOpts, EvmArgs, GlobalArgs},
     utils::{self, LoadConfig},
 };
-use foundry_common::{
-    EmptyTestFilter, TestFunctionExt, compile::ProjectCompiler, evm::EvmArgs, fs, shell,
-};
+use foundry_common::{EmptyTestFilter, TestFunctionExt, compile::ProjectCompiler, fs, shell};
 use foundry_compilers::{
     Language, ProjectCompileOutput, artifacts::output_selection::OutputSelection,
     compilers::multi::MultiCompiler, multi::MultiCompilerLanguage, utils::source_files_iter,
